P2의 경우, 서버로부터 공 위치를 받으면 공 위치를 바꾸는 기능 추가

+++ Assets/seunghoon/BallPositionModel.cs
@@ -5,11 +5,11 @@ namespace seunghoon
     [Serializable]
     public class BallPositionModel
     {
-        public double x;
-        public double y;
-        public double z;
+        public float x;
+        public float y;
+        public float z;
 
-        BallPositionModel(double x, double y, double z)
+        BallPositionModel(float x, float y, float z)
         {
             this.x = x;
             this.y = y;
